Move install_args_list_to_runnables to runnable_resolve
