+++ planemo/commands/cmd_shed_test.py
@@ -2,7 +2,6 @@
 import sys
 
 import click
-import requests
 
 from planemo import (
     options,
@@ -12,7 +11,7 @@ from planemo.cli import command_function
 from planemo.engine import engine_context
 from planemo.galaxy.test.actions import handle_reports_and_summary
 from planemo.network_util import get_free_port
-from planemo.runnable_resolve import for_runnable_identifier
+from planemo.runnable_resolve import install_args_list_to_runnables
 
 
 @click.command("shed_test")
@@ -41,20 +40,3 @@ def cli(ctx, paths, **kwds):
         ctx.vlog(f"engine.test returning [{test_data}]")
         return_code = handle_reports_and_summary(ctx, test_data.structured_data, kwds=kwds)
     sys.exit(return_code)
-
-
-def install_args_list_to_runnables(ctx, install_args_list, kwds):
-    runnables = []
-    for repo in install_args_list:
-        url = f'{repo["tool_shed_url"]}api/repositories/get_repository_revision_install_info'
-        response = requests.get(
-            url, params={"name": repo["name"], "owner": repo["owner"], "changeset_revision": repo["changeset_revision"]}
-        )
-        response.raise_for_status()
-        install_info = response.json()
-        repository_metadata = install_info[1]
-        assert repository_metadata["model_class"] == "RepositoryMetadata"
-        for tool in repository_metadata.get("valid_tools", []):
-            runnable = for_runnable_identifier(ctx, tool["guid"], kwds)
-            runnables.append(runnable)
-    return runnables
