+++ planemo/runnable_resolve.py
@@ -1,5 +1,7 @@
 import os
 
+import requests
+
 from planemo.galaxy.profiles import translate_alias
 from planemo.galaxy.workflows import GALAXY_WORKFLOWS_PREFIX
 from planemo.tools import uri_to_path
@@ -36,5 +38,22 @@ def for_runnable_identifiers(ctx, runnable_identifiers, kwds, temp_path=None):
         if isinstance(runnable, list):
             runnables.extend(runnable)
         else:
+            runnables.append(runnable)
+    return runnables
+
+
+def install_args_list_to_runnables(ctx, install_args_list, kwds):
+    runnables = []
+    for repo in install_args_list:
+        url = f'{repo["tool_shed_url"]}api/repositories/get_repository_revision_install_info'
+        response = requests.get(
+            url, params={"name": repo["name"], "owner": repo["owner"], "changeset_revision": repo["changeset_revision"]}
+        )
+        response.raise_for_status()
+        install_info = response.json()
+        repository_metadata = install_info[1]
+        assert repository_metadata["model_class"] == "RepositoryMetadata"
+        for tool in repository_metadata.get("valid_tools", []):
+            runnable = for_runnable_identifier(ctx, tool["guid"], kwds)
             runnables.append(runnable)
     return runnables
